[refactor/TK-215]: CookieUtil을 Security 내부로 이동
